requirements.txt
pytest==8.3.4
pytest-dotenv==0.5.2
python-dotenv==1.0.1
bcrypt==4.3.0

alembic==1.14.1
Mako==1.3.9


Flask==3.1.0
openai==1.65.1
psycopg2-binary==2.9.10
passlib==1.7.4
marshmallow==3.26.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1